requirements.txt
langchain==0.1.12
langchain-google-genai==0.0.11
google-generativeai>=0.4.1,<0.5.0
pandas==2.2.1
pydantic==2.6.4
python-dotenv==1.0.1
faiss-cpu==1.8.0
streamlit==1.32.2
sqlalchemy==2.0.28
python-multipart==0.0.9
numpy==1.26.4 